patch: improve latency for small ranges
Annoyingly we have to fight the compiler a bit to get this to work.

We "launch" from counting_sort, specifically including gaurds for low counts.
Unfortuantely, while it's possible to include these same if conditions inside the later
recursive calls up front, what will end up happening is they will be shuffled around such that
the intialization cost of all the book-keeping will be placed first.