performance: use small_merge_sort over insertion and heap sorts if possible
I'm assuming since counting sort already uses a lot of stack space we should be free to make use of more stack space in the failover case where possible.

This could backfire for large types. In testing with relatively normal objects like strings, string_views and integers it outperfrms insertion and heap sort for the entire range of 0->256.
The performance improvement at the lower end and the reduction of an if/else branch should be a performance impreovement overall.